+++ src/g_hexen/a_flechette.cpp
@@ -285,7 +285,7 @@ int APoisonCloud::DoSpecialDamage (AActor *victim, int damage)
 			int damage = 15 + (pr_poisoncloudd()&15);
 			if (mate)
 			{
-				damage = (int)((float)damage * level.teamdamage);
+				damage = (int)((double)damage * level.teamdamage);
 			}
 			if (damage > 0)
 			{
